+++ Exploration/ExplorationActor.php
@@ -9,17 +9,21 @@ use Heptacom\HeptaConnect\Core\Emission\Contract\EmissionActorInterface;
 use Heptacom\HeptaConnect\Core\Emission\Contract\EmitContextFactoryInterface;
 use Heptacom\HeptaConnect\Core\Emission\Contract\EmitterStackBuilderFactoryInterface;
 use Heptacom\HeptaConnect\Core\Exploration\Contract\ExplorationActorInterface;
-use Heptacom\HeptaConnect\Core\Mapping\Contract\MappingServiceInterface;
+use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\DependencyCollection;
+use Heptacom\HeptaConnect\Dataset\Base\TypedDatasetEntityCollection;
 use Heptacom\HeptaConnect\Portal\Base\Emission\Contract\EmitContextInterface;
 use Heptacom\HeptaConnect\Portal\Base\Emission\Contract\EmitterStackInterface;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExploreContextInterface;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerStackInterface;
-use Heptacom\HeptaConnect\Portal\Base\Mapping\Contract\MappingInterface;
+use Heptacom\HeptaConnect\Portal\Base\Mapping\MappedDatasetEntityStruct;
 use Heptacom\HeptaConnect\Portal\Base\Mapping\MappingComponentCollection;
 use Heptacom\HeptaConnect\Portal\Base\Mapping\MappingComponentStruct;
 use Heptacom\HeptaConnect\Portal\Base\Publication\Contract\PublisherInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
+use Heptacom\HeptaConnect\Storage\Base\Action\Mapping\Map\MappingMapPayload;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Mapping\MappingMapActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
 use Psr\Log\LoggerInterface;
 
@@ -31,8 +35,6 @@ class ExplorationActor implements ExplorationActorInterface
 
     private LoggerInterface $logger;
 
-    private MappingServiceInterface $mappingService;
-
     private EmissionActorInterface $emissionActor;
 
     private EmitContextFactoryInterface $emitContextFactory;
@@ -43,22 +45,24 @@ class ExplorationActor implements ExplorationActorInterface
 
     private StorageKeyGeneratorContract $storageKeyGenerator;
 
+    private MappingMapActionInterface $mappingMapAction;
+
     public function __construct(
         LoggerInterface $logger,
-        MappingServiceInterface $mappingService,
         EmissionActorInterface $emissionActor,
         EmitContextFactoryInterface $emitContextFactory,
         PublisherInterface $publisher,
         EmitterStackBuilderFactoryInterface $emitterStackBuilderFactory,
-        StorageKeyGeneratorContract $storageKeyGenerator
+        StorageKeyGeneratorContract $storageKeyGenerator,
+        MappingMapActionInterface $mappingMapAction
     ) {
         $this->logger = $logger;
-        $this->mappingService = $mappingService;
         $this->emissionActor = $emissionActor;
         $this->emitContextFactory = $emitContextFactory;
         $this->publisher = $publisher;
         $this->emitterStackBuilderFactory = $emitterStackBuilderFactory;
         $this->storageKeyGenerator = $storageKeyGenerator;
+        $this->mappingMapAction = $mappingMapAction;
     }
 
     public function performExploration(
@@ -186,8 +190,13 @@ class ExplorationActor implements ExplorationActorInterface
             \implode(',', $primaryKeys)
         ));
 
-        \iterable_filter($this->mappingService->getListByExternalIds($entityType, $portalNodeKey, $primaryKeys));
-        $this->emissionActor->performEmission($primaryKeys, clone $emissionStack, $emitContext);
+        $mapResult = $this->mappingMapAction->map(new MappingMapPayload($portalNodeKey, $this->factorizeMappableEntities($entityType, $primaryKeys)));
+
+        $this->emissionActor->performEmission(
+            $mapResult->getMappedDatasetEntityCollection()->map(static fn (MappedDatasetEntityStruct $me): ?string => $me->getMapping()->getExternalId()),
+            clone $emissionStack,
+            $emitContext
+        );
     }
 
     /**
@@ -207,19 +216,39 @@ class ExplorationActor implements ExplorationActorInterface
             \implode(',', $externalIds)
         ));
 
-        $this->publisher->publishBatch(new MappingComponentCollection($this->iterableValues(\iterable_map(
-            $this->mappingService->getListByExternalIds($entityType, $portalNodeKey, $externalIds),
-            static fn (MappingInterface $mapping): MappingComponentStruct => new MappingComponentStruct($portalNodeKey, $entityType, $mapping->getExternalId())
-        ))));
+        $entities = $this->factorizeMappableEntities($entityType, $externalIds);
+
+        $mapResult = $this->mappingMapAction->map(new MappingMapPayload($portalNodeKey, $entities));
+        $this->publisher->publishBatch(new MappingComponentCollection($mapResult->getMappedDatasetEntityCollection()->map(
+            static fn (MappedDatasetEntityStruct $me): MappingComponentStruct => new MappingComponentStruct(
+                $me->getMapping()->getPortalNodeKey(),
+                $me->getMapping()->getEntityType(),
+                $me->getMapping()->getExternalId()
+            )
+        )));
     }
 
     /**
-     * @TODO replace with iterable_values from bentools v2
+     * @param class-string<\Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract> $entityType
+     * @param string[]                                                                         $primaryKeys
      */
-    private function iterableValues(iterable $i): iterable
+    private function factorizeMappableEntities(string $entityType, array $primaryKeys): TypedDatasetEntityCollection
     {
-        foreach ($i as $item) {
-            yield $item;
+        $result = new TypedDatasetEntityCollection($entityType);
+        $entityFactory = new \ReflectionClass($entityType);
+
+        foreach ($primaryKeys as $primaryKey) {
+            /** @var DatasetEntityContract $entity */
+            $entity = $entityFactory->newInstanceWithoutConstructor();
+            \Closure::bind(function (DatasetEntityContract $entity): void {
+                $entity->attachments = new AttachmentCollection();
+                $entity->dependencies = new DependencyCollection();
+            }, null, $entity)($entity);
+            $entity->setPrimaryKey($primaryKey);
+
+            $result->push([$entity]);
         }
+
+        return $result;
     }
 }